third_party/blink/web_tests/http/tests/misc/image-blocked-src-no-change-expected.txt

CONSOLE ERROR: Not allowed to load local resource: compass.jpg
CONSOLE ERROR: Not allowed to load local resource: compass.jpg
Test case for bug 17897 : Not Rendering Images Imported from XHTML Document

This test checks that setting an image url should lead to load even if the image was in error and the url is not changed.

For this test to pass, you should see 2 messages in the console
